fix(next16): resolve build errors after codemod
- Move stale chapters PUT/DELETE handlers from collection route into
  the [chapterId] dynamic route where chapterId is actually defined
  (these were dead code, since the param doesn't exist on the parent
  route. Next 16's stricter route type-check now catches it).
- Update teacher-availability DELETE to use Promise<params> shape that
  Next 16 requires (codemod missed this handler).
- Replace JSX.Element with ReactElement in CourseDetailClient
  (React 19 dropped the global JSX namespace).
- Accept tsconfig.json changes Next 16 enforces:
  jsx: react-jsx, target: ES2017, .next/dev/types include.
